pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-wallet-py"
version = "0.1.0"
description = "Agentic Wallet Python Backend - Research-focused Web3 wallet explorer"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "httpx>=0.25.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "tenacity>=8.2.0",
    "structlog>=23.2.0",
    "python-multipart>=0.0.6",
    "eth-utils>=5.3.1",
    "anthropic>=0.8.0",
    "tiktoken>=0.5.0",
    "PyYAML>=6.0.1",
    "langgraph>=0.6.0",
    # Auth dependencies
    "siwe>=4.0.0",
    "PyNaCl>=1.5.0",
    "PyJWT>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.12.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["agentic_wallet_py*"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
target-version = "py311"
line-length = 100
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
